# prisma/swagger.yaml
openapi: "3.0.0"
paths:
  /api/car/shared:
    get:
      operationId: "CarAccessController_getSharedCars"
      summary: "Get all cars shared with the current user"
      parameters: []
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/SharedCarDto"
      tags:
        - "car-access"
      security:
        - cookie: []
  /api/car/{carId}/access:
    get:
      operationId: "CarAccessController_getAccessList"
      summary: "List all users with access to a car"
      parameters:
        - name: "carId"
          required: true
          in: "path"
          schema:
            type: "number"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/CarAccessDto"
      tags:
        - "car-access"
      security:
        - cookie: []
  /api/car/{carId}/access/invite:
    post:
      operationId: "CarAccessController_inviteUser"
      summary: "Invite a registered user to access a car"
      parameters:
        - name: "carId"
          required: true
          in: "path"
          schema:
            type: "number"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteUserDto"
      responses:
        201:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarAccessDto"
      tags:
        - "car-access"
      security:
        - cookie: []
  /api/car/{carId}/access/accept:
    post:
      operationId: "CarAccessController_acceptInvitation"
      summary: "Accept a pending invitation to a car"
      parameters:
        - name: "carId"
          required: true
          in: "path"
          schema:
            type: "number"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarAccessDto"
      tags:
        - "car-access"
      security:
        - cookie: []
  /api/car/{carId}/access/{targetUserId}/role:
    patch:
      operationId: "CarAccessController_changeRole"
      summary: "Change the role of a user on a car"
      parameters:
        - name: "carId"
          required: true
          in: "path"
          schema:
            type: "number"
        - name: "targetUserId"
          required: true
          in: "path"
          schema:
            type: "number"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeRoleDto"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarAccessDto"
      tags:
        - "car-access"
      security:
        - cookie: []
  /api/car/{carId}/access/leave:
    delete:
      operationId: "CarAccessController_leaveAccess"
      summary: "Leave (remove yourself from) a shared car"
      parameters:
        - name: "carId"
          required: true
          in: "path"
          schema:
            type: "number"
      responses:
        204:
          description: "Access removed"
      tags:
        - "car-access"
      security:
        - cookie: []
  /api/car/{carId}/access/{targetUserId}:
    delete:
      operationId: "CarAccessController_removeAccess"
      summary: "Remove a user's access to a car"
      parameters:
        - name: "carId"
          required: true
          in: "path"
          schema:
            type: "number"
        - name: "targetUserId"
          required: true
          in: "path"
          schema:
            type: "number"
      responses:
        204:
          description: "Access removed"
      tags:
        - "car-access"
      security:
        - cookie: []
  /api/car:
    post:
      operationId: "CarController_createCar"
      summary: "Create a new car"
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/AddCarDto"
      responses:
        201:
          description: "Car created successfully."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarDto"
        403:
          description: "Forbidden."
      tags:
        - "car"
      security:
        - cookie: []
    put:
      operationId: "CarController_updateCar"
      summary: "Update a car"
      parameters: []
      responses:
        200:
          description: "Car updated successfully."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarDto"
        404:
          description: "Car not found."
      tags:
        - "car"
      security:
        - cookie: []
    get:
      operationId: "CarController_getAllCars"
      summary: "Get all cars for authenticated user"
      parameters: []
      responses:
        200:
          description: "Return list of cars."
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/CarDto"
      tags:
        - "car"
      security:
        - cookie: []
  /api/car/{id}:
    get:
      operationId: "CarController_getCar"
      summary: "Get car by ID"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: "Return car details."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarDto"
        404:
          description: "Car not found."
      tags:
        - "car"
      security:
        - cookie: []
    delete:
      operationId: "CarController_deleteCar"
      summary: "Delete a car"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: "Car deleted successfully."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarDto"
        404:
          description: "Car not found."
      tags:
        - "car"
      security:
        - cookie: []
  /api/car/photo/{photoId}:
    delete:
      operationId: "CarController_deletePhoto"
      summary: "Delete a car photo"
      parameters:
        - name: "photoId"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        204:
          description: "Photo deleted."
        404:
          description: "Photo not found."
      tags:
        - "car"
      security:
        - cookie: []
  /api/car/photo/{photoId}/default:
    patch:
      operationId: "CarController_setDefaultPhoto"
      summary: "Set a photo as the default for its car"
      parameters:
        - name: "photoId"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        204:
          description: "Default photo updated."
        404:
          description: "Photo not found."
      tags:
        - "car"
      security:
        - cookie: []
  /api/car/{id}/sold:
    patch:
      operationId: "CarController_markAsSold"
      summary: "Mark a car as sold (archive)"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: "Car marked as sold."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarDto"
        403:
          description: "Only the owner can archive this car."
      tags:
        - "car"
      security:
        - cookie: []
  /api/car/{id}/restore:
    patch:
      operationId: "CarController_restoreCar"
      summary: "Restore a sold car back to active"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: "Car restored to active."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarDto"
        403:
          description: "Only the owner can restore this car."
      tags:
        - "car"
      security:
        - cookie: []
  /api/car/user/{userId}:
    get:
      operationId: "CarController_getCarsByUser"
      summary: "Get all cars for a specific user"
      parameters:
        - name: "userId"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: "Return list of cars."
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/CarDto"
      tags:
        - "car"
      security:
        - cookie: []
...
  /api/maintenance-record/car/{carId}:
    get:
      operationId: "MaintenanceRecordController_getMaintenanceRecordsByCarId"
      summary: "Get all maintenance records for a specific car"
      parameters:
        - name: "carId"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: "Return list of maintenance records."
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/MaintenanceRecordDto"
      tags:
        - "maintenance-record"
      security:
        - bearer: []
...
  /api/document/car/{carId}:
    get:
      operationId: "DocumentController_getDocumentsByCarId"
      summary: "Get all documents for a specific car"
      parameters:
        - name: "carId"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/DocumentDto"
      tags:
        - "document"
...
info:
  title: "History Auto Utility API"
  description: "API documentation for History Auto Utility"
  version: "1.0"
  contact: {}
...
servers:
  - url: "http://localhost:3000"
    description: "development"
components:
  securitySchemes:
    bearer:
      scheme: "bearer"
      bearerFormat: "JWT"
      type: "http"
  schemas:
    SharedCarDto:
      type: "object"
      properties:
        id:
          type: "number"
        make:
          type: "string"
        model:
          type: "string"
        variant:
          type: "string"
        year:
          type: "number"
        license_plate:
          type: "string"
        shared_role:
          type: "string"
          enum:
            - "OWNER"
            - "FULL"
            - "USER"
            - "MAINTENANCE"
            - "VIEWER"
        accepted_at:
          format: "date-time"
          type: "string"
      required:
        - "id"
        - "make"
        - "model"
        - "year"
        - "license_plate"
        - "shared_role"
    CarAccessUserDto:
      type: "object"
      properties:
        id:
          type: "number"
        email:
          type: "string"
        first_name:
          type: "string"
        last_name:
          type: "string"
        picture:
          type: "string"
      required:
        - "id"
        - "email"
        - "first_name"
        - "last_name"
    CarAccessDto:
      type: "object"
      properties:
        id:
          type: "number"
        car_id:
          type: "number"
        role:
          type: "string"
          enum:
            - "OWNER"
            - "FULL"
            - "USER"
            - "MAINTENANCE"
            - "VIEWER"
        user:
          $ref: "#/components/schemas/CarAccessUserDto"
        invited_by:
          $ref: "#/components/schemas/CarAccessUserDto"
        accepted_at:
          format: "date-time"
          type: "string"
        created_at:
          format: "date-time"
          type: "string"
      required:
        - "id"
        - "car_id"
        - "role"
        - "user"
        - "invited_by"
        - "created_at"
    InviteUserDto:
      type: "object"
      properties:
        email:
          type: "string"
          example: "friend@example.com"
        role:
          type: "string"
          enum:
            - "OWNER"
            - "FULL"
            - "USER"
            - "MAINTENANCE"
            - "VIEWER"
          example: "VIEWER"
      required:
        - "email"
        - "role"
    ChangeRoleDto:
      type: "object"
      properties:
        role:
          type: "string"
          enum:
            - "OWNER"
            - "FULL"
            - "USER"
            - "MAINTENANCE"
            - "VIEWER"
          example: "VIEWER"
      required:
        - "role"
    AddCarDto:
      type: "object"
      properties:
        id:
          type: "number"
          example: "6"
        make:
          type: "string"
          example: "BMW"
        model:
          type: "string"
          example: "520i"
        variant:
          type: "string"
          example: "M Sport"
        license_plate:
          type: "string"
          example: "IS49WAU"
        vin:
          type: "string"
          example: "WBAHE21060GE64612"
        nickname:
          type: "string"
          example: "Blue Beast"
        year:
          type: "number"
          example: 2006
        fuel_type:
          type: "string"
          example: "PETROL"
          enum:
            - "PETROL"
            - "DIESEL"
            - "HYBRID"
            - "PLUGIN_HYBRID"
            - "ELECTRIC"
            - "LPG"
        transmission:
          type: "string"
          example: "MANUAL"
          enum:
            - "MANUAL"
            - "AUTOMATIC"
            - "SEMI_AUTOMATIC"
        engine:
          type: "string"
          example: "2.0L Turbo"
        color:
          type: "string"
          example: "#1A2B3C"
        current_mileage:
          type: "number"
          example: 184200
        ownership_start_date:
          type: "string"
          example: "2024-05-12T00:00:00.000Z"
        rca_expiry_date:
          type: "string"
          example: "2026-03-10T00:00:00.000Z"
        itp_expiry_date:
          type: "string"
          example: "2026-04-20T00:00:00.000Z"
        rov_expiry_date:
          type: "string"
          example: "2026-05-15T00:00:00.000Z"
        last_oil_service_date:
          type: "string"
          example: "2026-03-12T00:00:00.000Z"
        last_oil_service_mileage:
          type: "number"
          example: 180000
        default_new_photo_index:
          type: "number"
          example: 0
      required:
        - "id"
        - "make"
        - "model"
        - "variant"
        - "vin"
        - "nickname"
        - "year"
        - "fuel_type"
        - "transmission"
        - "engine"
        - "color"
        - "ownership_start_date"
        - "rca_expiry_date"
        - "itp_expiry_date"
        - "rov_expiry_date"
        - "last_oil_service_date"
        - "last_oil_service_mileage"
        - "default_new_photo_index"
    CarPhotoDto:
      type: "object"
      properties:
        id:
          type: "number"
          example: 1
        url:
          type: "string"
          example: "/uploads/cars/photo.jpg"
        is_default:
          type: "boolean"
          example: false
      required:
        - "id"
        - "url"
        - "is_default"
    CarDto:
      type: "object"
      properties:
        id:
          type: "number"
          example: 1
        user_id:
          type: "number"
          example: 123
        vin:
          type: "string"
          example: "1HGCM82633A123456"
        nickname:
          type: "string"
          example: "Blue Beast"
        make:
          type: "string"
          example: "BMW"
        model:
          type: "string"
          example: "530iA"
        variant:
          type: "string"
          example: "M Sport"
        year:
          type: "number"
          example: 1992
        license_plate:
          type: "string"
          example: "IS49WAU"
        fuel_type:
          type: "string"
          example: "PETROL"
          enum:
            - "PETROL"
            - "DIESEL"
            - "HYBRID"
            - "PLUGIN_HYBRID"
            - "ELECTRIC"
            - "LPG"
        transmission:
          type: "string"
          example: "MANUAL"
          enum:
            - "MANUAL"
            - "AUTOMATIC"
            - "SEMI_AUTOMATIC"
        engine:
          type: "string"
          example: "2.0L Turbo"
        color:
          type: "string"
          example: "#1A2B3C"
        current_mileage:
          type: "number"
          example: 184200
        ownership_start_date:
          format: "date-time"
          type: "string"
          example: "2024-05-12T00:00:00.000Z"
        rca_expiry_date:
          format: "date-time"
          type: "string"
          example: "2026-03-10T00:00:00.000Z"
        itp_expiry_date:
          format: "date-time"
          type: "string"
          example: "2026-04-20T00:00:00.000Z"
        rov_expiry_date:
          format: "date-time"
          type: "string"
          example: "2026-05-15T00:00:00.000Z"
        last_oil_service_date:
          format: "date-time"
          type: "string"
          example: "2026-03-12T00:00:00.000Z"
        last_oil_service_mileage:
          type: "number"
          example: 180000
        status:
          type: "string"
          example: "ACTIVE"
          enum:
            - "ACTIVE"
            - "SOLD"
        sold_at:
          format: "date-time"
          type: "string"
          example: "2025-06-01T00:00:00.000Z"
        photos:
          type: "array"
          items:
            $ref: "#/components/schemas/CarPhotoDto"
      required:
        - "id"
        - "user_id"
        - "vin"
        - "nickname"
        - "make"
        - "model"
        - "variant"
        - "year"
        - "license_plate"
        - "fuel_type"
        - "transmission"
        - "engine"
        - "color"
        - "ownership_start_date"
        - "rca_expiry_date"
        - "itp_expiry_date"
        - "rov_expiry_date"
        - "last_oil_service_date"
        - "last_oil_service_mileage"
        - "status"
        - "photos"
...
    MaintenanceRecordDto:
      type: "object"
      properties:
        id:
          type: "number"
          example: "123"
        car_id:
          type: "number"
          example: "123"
        service_date:
          format: "date-time"
          type: "string"
          example: "2023-01-01T00:00:00.000Z"
        mileage:
          type: "number"
          example: "380000"
        description:
          type: "string"
          example: "Oil 10W40, air filter MANN"
        service_type:
          type: "string"
          example: "MAINTENANCE"
          enum:
            - "REPAIR"
            - "MAINTENANCE"
            - "IMPROVEMENT"
        service_category:
          type: "string"
          example: "OIL_CHANGE"
          enum:
            - "OIL_CHANGE"
            - "BRAKE_SERVICE"
            - "TRANSMISSION_SERVICE"
            - "TIRE_SERVICE"
            - "FLUID_SERVICE"
            - "ENGINE_SERVICE"
            - "INSPECTION"
            - "BATTERY_SERVICE"
            - "FILTER_SERVICE"
            - "LIGHT_SERVICE"
            - "OTHER"
        cost:
          type: "number"
          example: "Insurance"
        expiry_date:
          format: "date-time"
          type: "string"
          example: "Insurance"
      required:
        - "id"
        - "car_id"
        - "service_date"
        - "mileage"
        - "description"
        - "service_type"
        - "service_category"
        - "cost"
        - "expiry_date"
...
    DocumentDto:
      type: "object"
      properties:
        id:
          type: "number"
          example: 1
        document_type:
          type: "string"
          example: "RCA"
        car_id:
          type: "number"
          example: 1
        issue_date:
          format: "date-time"
          type: "string"
          nullable: true
        expiry_date:
          format: "date-time"
          type: "string"
          nullable: true
        provider:
          type: "string"
          nullable: true
        policy_number:
          type: "string"
          nullable: true
        status:
          type: "string"
          nullable: true
        policyholder:
          type: "string"
          nullable: true
        cnp_id:
          type: "string"
          nullable: true
        file_url:
          type: "string"
          nullable: true
        file_name:
          type: "string"
          nullable: true
        file_size:
          type: "number"
          nullable: true
      required:
        - "id"
        - "document_type"
        - "car_id"
...